requirements.txt
opencv-python
pytest
python-dotenv